fixed mistake made when pushing Mesh/TextureData to 64 bytes, cause indirection + cleanup
cleaned up all warnings that I know how to deal with.
made MeshData 80 bytes, to provide all data iterative data, simiarly TextureData is now 72 byes to provide all iterative data. All rarely accessed data is hidden away by a shared ptr, for ease of ptr managenment.

+++ IciclEngine/include/engine/resources/data_storage.h
@@ -126,7 +126,7 @@ protected:
 	bool start = false;
 	std::condition_variable cv_start_threads;
 	std::vector<std::thread> threads;
-	int num_threads;
+	size_t num_threads;
 };
 
 struct ModelGenStorage : GenStorageThreadPool // this will have more job processors, but only one genstoragethreadpool
